+++ src/plugins/heartFC_chat/heartFC_processor.py
