+++ src/plugins/heartFC_chat/pfchating.md
@@ -93,8 +93,5 @@
 
 
 BUG:
-1.第一条激活消息没有被读取，进入pfc聊天委托时应该读取一下之前的上文(fix)
 2.复读，可能是planner还未校准好
 3.planner还未个性化，需要加入bot个性信息，且获取的聊天内容有问题
-4.心流好像过短，而且有时候没有等待更新
-5.表情包有可能会发两次(fix)
