+++ src/plugins/heartFC_chat/reasoning_chat.py
